+++ src/Indice.Features.Cases.AspNetCore/Services/AdminCaseService.cs
@@ -614,6 +614,18 @@ internal class AdminCaseService : BaseCaseService, IAdminCaseService
         return timeline;
     }
 
+    public async Task<List<CasePartial>> GetRelatedCases(ClaimsPrincipal user, Guid caseId) {
+        // Check that user role can view this case
+        var @case = await GetCaseById(user, caseId, false);
+        var result = await GetCases(user, new ListOptions<GetCasesListFilter>() {
+            Filter = new GetCasesListFilter {
+                Metadata = [new FilterClause("metadata.ExternalCorrelationKey", @case.Metadata["ExternalCorrelationKey"], FilterOperator.Eq, JsonDataType.String)]
+            }
+        });
+
+        return result.Items.OrderByDescending(x=> x.CreatedByWhen).ToList();
+    }
+
     private async Task<List<FilterClause>> MapCheckpointTypeCodeToId(List<FilterClause> checkpointTypeCodeFilterClauses) {
         var checkpointTypeCodes = checkpointTypeCodeFilterClauses.Select(f => f.Value).ToList();
         var checkpointTypeIds = new List<FilterClause>();