+++ src/Indice.Features.Media.AspNetCore/Endpoints/MediaSettingsApi.cs
@@ -15,7 +15,7 @@ public static class MediaSettingsApi
     /// <returns>The <see cref="IEndpointRouteBuilder"/> instance.</returns>
     internal static IEndpointRouteBuilder MapMediaSettings(this IEndpointRouteBuilder builder) {
         var options = builder.ServiceProvider.GetService<IOptions<MediaApiOptions>>()?.Value ?? new MediaApiOptions();
-        var group = builder.MapGroup($"{options.ApiPrefix}/media/settings")
+        var group = builder.MapGroup($"{options.PathPrefix.Value!.TrimEnd('/')}/media/settings")
                            .WithGroupName("media")
                            .WithTags("Settings")
                            .ProducesProblem(StatusCodes.Status401Unauthorized)
@@ -24,7 +24,7 @@ public static class MediaSettingsApi
                            .RequireAuthorization(MediaLibraryApi.Policies.BeMediaLibraryManager)
                            .WithHandledException<BusinessException>();
 
-        var requiredScopes = options.ApiScope.Split(' ').Where(scope => !string.IsNullOrWhiteSpace(scope)).ToArray();
+        var requiredScopes = options.Scope.Split(' ').Where(scope => !string.IsNullOrWhiteSpace(scope)).ToArray();
         group.WithOpenApi().AddOpenApiSecurityRequirement("oauth2", requiredScopes);
 
         group.MapGet("", MediaSettingsHandlers.ListMediaSettings)