+++ src/Indice.Features.Messages.Core/Services/CampaignAttachmentService.cs
@@ -43,7 +43,7 @@ public class CampaignAttachmentService : ICampaignAttachmentService
             ContentType = fileAttachment.ContentType,
             Id = attachment.Id,
             Label = fileAttachment.Name,
-            PermaLink = $"{CampaignManagementOptions.ApiPrefix}/campaigns/attachments/{(Base64Id)attachment.Guid}.{Path.GetExtension(attachment.Name).TrimStart('.')}",
+            PermaLink = $"{CampaignManagementOptions.PathPrefix}/campaigns/attachments/{(Base64Id)attachment.Guid}.{Path.GetExtension(attachment.Name).TrimStart('.')}",
             Size = fileAttachment.ContentLength
         };
     }