+++ src/Indice.Features.Messages.UI/CampaignUIOptions.cs
@@ -11,13 +11,20 @@ public class CampaignUIOptions : SpaUIOptions
     /// <summary>Enables a rich text editor for Campaign content.</summary>
     public bool EnableRichTextEditor { get; set; } = false;
 #endif
+    /// <summary>Enables a rich text editor for Campaign content. defaults to <strong>/api</strong> which is the same origin. </summary>
+    /// <remarks>Can also be pointed to a an exteranl origin ei: messaging.indice.gr.</remarks>
+    public string ApiBaseUrl { get; set; } = "/api";
     /// <summary>Creates a new instance <see cref="CampaignUIOptions"/>.</summary>
     public CampaignUIOptions() {
+        ClientId = "messaging-ui";
+        Scope = "offline_access messages media";
+        DocumentTitle = "Messaging UI";
         ConfigureIndexParameters = (args) => {
 #if NET7_0_OR_GREATER
             args.Add($"%({nameof(EnableMediaLibrary)})", EnableMediaLibrary.ToString());
             args.Add($"%({nameof(EnableRichTextEditor)})", EnableRichTextEditor.ToString());
 #endif
+            args.Add($"%({nameof(ApiBaseUrl)})", ApiBaseUrl);
         };
     }
 }