+++ test/Indice.Features.Messages.Tests/MessagesIntegrationTests.cs
@@ -1,4 +1,4 @@
-﻿using System;
+﻿#if NET7_0_OR_GREATER
 using System.Text;
 using System.Text.Json;
 using Indice.AspNetCore.Authorization;
@@ -44,9 +44,9 @@ public class MessagesIntegrationTests : IAsyncLifetime
         builder.ConfigureServices(services => {
             var configuration = services.BuildServiceProvider().GetService<IConfiguration>();
             services.AddTransient<IEventDispatcherFactory, DefaultEventDispatcherFactory>();
-            services.AddControllers()
-                    .AddMessageEndpoints(options => {
-                        options.ApiPrefix = "api";
+            services.AddRouting();
+            services.AddMessaging(options => {
+                        options.PathPrefix = "api";
                         options.ConfigureDbContext = (serviceProvider, dbbuilder) => dbbuilder.UseSqlServer(configuration.GetConnectionString("MessagesDb"));
                         options.DatabaseSchema = "cmp";
                         options.RequiredScope = MessagesApi.Scope;
@@ -64,7 +64,7 @@ public class MessagesIntegrationTests : IAsyncLifetime
             app.UseAuthentication();
             app.UseRouting();
             app.UseAuthorization();
-            app.UseEndpoints(e => e.MapControllers());
+            app.UseEndpoints(e => e.MapMessaging());
         });
         var server = new TestServer(builder);
         var handler = server.CreateHandler();
@@ -73,7 +73,6 @@ public class MessagesIntegrationTests : IAsyncLifetime
         };
     }
 
-    #region Facts
     [Fact]
     public async Task Create_And_Retrieve_Campaign_By_Location_Header_Success() {
         //Create the Campaign
@@ -84,7 +83,7 @@ public class MessagesIntegrationTests : IAsyncLifetime
                 To = DateTimeOffset.UtcNow.AddDays(1)
             },
             Published = false,
-            RecipientIds = new List<string> { "6c9fa6dd-ede4-486b-bf91-6de18542da4a" },
+            RecipientIds = [ "6c9fa6dd-ede4-486b-bf91-6de18542da4a" ],
             Content = new MessageContentDictionary(
                 new Dictionary<MessageChannelKind, MessageContent> {
                     [MessageChannelKind.Inbox] = new MessageContent("Test Message", "Test Message Content")
@@ -204,7 +203,6 @@ public class MessagesIntegrationTests : IAsyncLifetime
         Assert.True(createTemplateResponse.IsSuccessStatusCode);
         Assert.True(getTemplateResponse.IsSuccessStatusCode);
     }
-    #endregion
 
     public async Task InitializeAsync() {
         var db = _serviceProvider.GetRequiredService<CampaignsDbContext>();
@@ -217,3 +215,5 @@ public class MessagesIntegrationTests : IAsyncLifetime
         await _serviceProvider.DisposeAsync();
     }
 }
+
+#endif